minor optimisation: replaced a int with a byte in WayDescription
I don't expect wayWidth to ever be bigger than 255, so no need for an int.
Dosn't save much though, as there are only a few WayDescription objects

+++ GpsMidGraph/de/ueller/midlet/gps/tile/WayDescription.java
@@ -7,6 +7,6 @@ public class WayDescription {
 	public int lineStyle;
 	public int boardedColor;
 	public boolean isArea;
-	public int wayWidth;
+	public byte wayWidth;
 
 }
